# documents/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User API",
        "description": "API para gestionar usuarios en el sistema",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "141957515afad3088dce4d73ca0e36cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sandra Montero"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "sandra@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "User created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Iniciar sesión",
                "operationId": "ad685b02e9c4cac1c1f4dc8268e3a3de",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "sandra1@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "properties": {
                                                "headers": {
                                                    "type": "object",
                                                    "example": []
                                                },
                                                "original": {
                                                    "properties": {
                                                        "user": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 6
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Sandra"
                                                                },
                                                                "email": {
                                                                    "type": "string",
                                                                    "example": "sandra1@gmail.com"
                                                                },
                                                                "email_verified_at": {
                                                                    "type": "string",
                                                                    "example": null,
                                                                    "nullable": true
                                                                },
                                                                "created_at": {
                                                                    "type": "string",
                                                                    "format": "date-time",
                                                                    "example": "2024-12-07T17:29:11.000000Z"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string",
                                                                    "format": "date-time",
                                                                    "example": "2024-12-07T17:29:11.000000Z"
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "token": {
                                                            "type": "string",
                                                            "example": "1|TMmazWlwprJsmI6Mqqav36g8pfA7oIFyuB2zLRP6d2a5a7df"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "exception": {
                                                    "type": "object",
                                                    "example": null,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciales inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Obtener un usuario por ID",
                "operationId": "a3b180f7f8d27b5cf0389021ad008a03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "User"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "name": "Admin User",
                                                "email": "admin1@example.com",
                                                "emailVerifiedDate": ""
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Actualizar un usuario",
                "operationId": "952fa07e3c800c88662a1d37d8ea9845",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sandra Montero Sanz"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "sandra1@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "User updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error al actualizar el usuario"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "dc0ac7b984bd20006f08a9e4c526bf72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "User deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}